Move CLI utility classes in leshan-core-demo to be reused by other demos
