+++ leshan-client-demo/logback-config.xml
@@ -15,9 +15,9 @@ Contributors:
     Sierra Wireless - initial API and implementation
  -->
 <configuration>
-	<appender name="TERMINAL" class="org.eclipse.leshan.client.demo.cli.interactive.TerminalAppender">
+	<appender name="TERMINAL" class="org.eclipse.leshan.core.demo.cli.interactive.TerminalAppender">
 		<encoder class="ch.qos.logback.core.encoder.LayoutWrappingEncoder">
-			<layout class="org.eclipse.leshan.client.demo.logback.ColorAwarePatternLayout">
+			<layout class="org.eclipse.leshan.core.demo.logback.ColorAwarePatternLayout">
 				<pattern>%gray(%30.30logger{0}) %gray(%d) [%highlight(%p)] %m%n</pattern>
 			</layout>
 		</encoder>
