+++ leshan-client-demo/pom.xml
@@ -35,18 +35,6 @@ Contributors:
             <groupId>org.eclipse.leshan</groupId>
             <artifactId>leshan-core-demo</artifactId>
         </dependency>
-        <dependency>
-            <groupId>info.picocli</groupId>
-            <artifactId>picocli</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>info.picocli</groupId>
-            <artifactId>picocli-shell-jline2</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>ch.qos.logback</groupId>
-            <artifactId>logback-classic</artifactId>
-        </dependency>
     </dependencies>
 
     <build>
