+++ leshan-core-demo/src/main/java/org/eclipse/leshan/core/demo/cli/ExecutionExceptionHandler.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.client.demo.cli;
+package org.eclipse.leshan.core.demo.cli;
 
 import java.io.PrintWriter;
