+++ leshan-core-demo/src/main/java/org/eclipse/leshan/core/demo/cli/converters/PortConverter.java
@@ -1,0 +1,32 @@
+/*******************************************************************************
+ * Copyright (c) 2021 Sierra Wireless and others.
+ * 
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * and Eclipse Distribution License v1.0 which accompany this distribution.
+ * 
+ * The Eclipse Public License is available at
+ *    http://www.eclipse.org/legal/epl-v20.html
+ * and the Eclipse Distribution License is available at
+ *    http://www.eclipse.org/org/documents/edl-v10.html.
+ * 
+ * Contributors:
+ *     Sierra Wireless - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.leshan.core.demo.cli.converters;
+
+import picocli.CommandLine.ITypeConverter;
+
+public class PortConverter implements ITypeConverter<Integer> {
+
+    @Override
+    public Integer convert(String value) throws Exception {
+        Integer res = Integer.parseInt(value);
+
+        if (res < 0 || 65535 < res)
+            throw new IllegalArgumentException(
+                    String.format("%s is not a valid port number. It should be between 0 and 65535", value));
+
+        return res;
+    }
+}
