+++ leshan-core-demo/src/main/java/org/eclipse/leshan/core/demo/logback/ColorAwarePatternLayout.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.client.demo.logback;
+package org.eclipse.leshan.core.demo.logback;
 
 import ch.qos.logback.classic.PatternLayout;
 import picocli.CommandLine.Help.Ansi;
